+++ pueue_lib/src/process_helper/mod.rs
@@ -22,6 +22,7 @@ use command_group::Signal;
 #[cfg_attr(target_vendor = "apple", path = "apple.rs")]
 #[cfg_attr(target_os = "windows", path = "windows.rs")]
 #[cfg_attr(target_os = "freebsd", path = "freebsd.rs")]
+#[cfg_attr(target_os = "netbsd", path = "netbsd.rs")]
 mod platform;
 pub use self::platform::*;